Added stfl library.
Since it will be removed from the portage, because of the breakage on
the latest python version, I've added "new" ebuild, which removes all
the bindings from it, just to not have to deal with it in the future.
The only application which is depending on this lib is newsboat, which
as far as I can tell, doesn't use the bindings.